feat(speakers): run the speaker import on the pg-boss queue
The import endpoint writes the upload to the job-files volume and enqueues a
job carrying its path, replacing the unawaited promise and the in-memory
`Map` that no container restart could survive. Job state is now pg-boss
state: the status endpoint reads it with `findJobs` and maps it onto the four
states the client already understands, presenting a retry window as
`processing` so a backoff does not read as a failure.

The 20 MB size guard and the 1.5 s client polling are unchanged.

Refs: GRX-4, GRX-3
